# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal Digital Twin Assistant API",
    "description": "Backend API for the Personal Digital Twin productivity assistant",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/profile": {
      "get": {
        "tags": [
          "Profile Memory"
        ],
        "summary": "Get Profile Memory",
        "description": "Return active long-term memory and recent extraction events.",
        "operationId": "get_profile_memory_api_profile_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileMemoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/profile/context": {
      "post": {
        "tags": [
          "Profile Memory"
        ],
        "summary": "Create Context Rule",
        "description": "Save a user-approved AI memory fact.",
        "operationId": "create_context_rule_api_profile_context_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserContextCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/profile/questions": {
      "get": {
        "tags": [
          "Profile Memory"
        ],
        "summary": "Get Profile Questions",
        "description": "Return a short guided profile questionnaire.",
        "operationId": "get_profile_questions_api_profile_questions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProfileQuestionResponse"
                  },
                  "type": "array",
                  "title": "Response Get Profile Questions Api Profile Questions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Profile Memory"
        ],
        "summary": "Submit Profile Questions",
        "description": "Save reviewed questionnaire answers as explicit user-approved memory.",
        "operationId": "submit_profile_questions_api_profile_questions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileQuestionnaireSubmit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserContextResponse"
                  },
                  "type": "array",
                  "title": "Response Submit Profile Questions Api Profile Questions Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/profile/context/{context_id}": {
      "put": {
        "tags": [
          "Profile Memory"
        ],
        "summary": "Update Context Rule",
        "description": "Correct an active AI memory fact.",
        "operationId": "update_context_rule_api_profile_context__context_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Context Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserContextUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Profile Memory"
        ],
        "summary": "Delete Context Rule",
        "description": "Remove an active AI memory fact.",
        "operationId": "delete_context_rule_api_profile_context__context_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Context Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProfileQuestionAnswer": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "category": {
            "type": "string",
            "enum": [
              "identity",
              "constraint",
              "goal",
              "general"
            ],
            "title": "Category"
          },
          "fact": {
            "type": "string",
            "minLength": 1,
            "title": "Fact"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "category",
          "fact"
        ],
        "title": "ProfileQuestionAnswer",
        "description": "Reviewed questionnaire answer that can be stored as memory."
      },
      "ProfileQuestionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "enum": [
              "identity",
              "constraint",
              "goal",
              "general"
            ],
            "title": "Category"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "placeholder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placeholder"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "prompt"
        ],
        "title": "ProfileQuestionResponse",
        "description": "A guided question used to collect user-approved profile memory."
      },
      "ProfileQuestionnaireSubmit": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/ProfileQuestionAnswer"
            },
            "type": "array",
            "title": "Answers"
          }
        },
        "type": "object",
        "title": "ProfileQuestionnaireSubmit",
        "description": "Batch payload for saving reviewed questionnaire answers."
      },
      "ProfileUpdateResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "enum": [
              "identity",
              "constraint",
              "goal",
              "general"
            ],
            "title": "Category"
          },
          "fact": {
            "type": "string",
            "title": "Fact"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "fact"
        ],
        "title": "ProfileUpdateResponse",
        "description": "Append-only memory extraction event shown in the profile feed."
      },
...
      "UserContextCreate": {
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "identity",
              "constraint",
              "goal",
              "general"
            ],
            "title": "Category"
          },
          "fact": {
            "type": "string",
            "minLength": 1,
            "title": "Fact"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "default": "user"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 1.0
          }
        },
        "type": "object",
        "required": [
          "category",
          "fact"
        ],
        "title": "UserContextCreate",
        "description": "Payload for explicitly saving a new long-term memory fact."
      },
      "UserContextResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "enum": [
              "identity",
              "constraint",
              "goal",
              "general"
            ],
            "title": "Category"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "fact": {
            "type": "string",
            "title": "Fact"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "key",
          "fact"
        ],
        "title": "UserContextResponse",
        "description": "Current long-term memory fact shown in the profile UI."
      },
      "UserContextUpdate": {
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "identity",
              "constraint",
              "goal",
              "general"
            ],
            "title": "Category"
          },
          "fact": {
            "type": "string",
            "minLength": 1,
            "title": "Fact"
          }
        },
        "type": "object",
        "required": [
          "category",
          "fact"
        ],
        "title": "UserContextUpdate",
        "description": "Payload for correcting an active long-term memory fact."
      },
...
      "UserProfileMemoryResponse": {
        "properties": {
          "context": {
            "items": {
              "$ref": "#/components/schemas/UserContextResponse"
            },
            "type": "array",
            "title": "Context"
          },
          "updates": {
            "items": {
              "$ref": "#/components/schemas/ProfileUpdateResponse"
            },
            "type": "array",
            "title": "Updates"
          }
        },
        "type": "object",
        "title": "UserProfileMemoryResponse",
        "description": "Current profile memory plus recent extraction log."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}